made changes  on strategies
